% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_quiz.R
\name{parse_quiz}
\alias{parse_quiz}
\title{Creates a fully formatted moodle quiz dataset}
\usage{
parse_quiz(question_stats, question_xml, response,
  question_change = NULL)
}
\arguments{
\item{question_stats}{Output from readr::read_csv() applied to Moodle Quiz structure analysis CSV file. From the quiz settings, select "Results" then "Statistics" then download "Quiz structure analysis" table as CSV file.}

\item{question_xml}{Output from xml2::read_xml() applied to Moodle Quiz XML file. From the quiz settings, select "Question Bank" then "Export" then download "Moodle XML" file.}

\item{response}{Output from readr::read_csv() applied to Moodle Quiz responses CSV file. From the quiz settings, select "Results" then "Responses" then download table as CSV file.}

\item{question_change}{Default = NULL. Not recommended, but can be used if a question is added after a quiz goes live. Must supply a tibble containing 3 columns: "response" (what question in the dataset that has been added), "change" (the change that has been made - only "new" questions accounted for at present), and "date" (the date the change was made).}
}
\value{
Returns a fully formatted dataset for the moodle quiz
}
\description{
Wrangles files directly extracted from moodle to create a fully formatted dataset for a moodle quiz
}
