% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_quiz_metadata.R
\name{parse_quiz_metadata}
\alias{parse_quiz_metadata}
\title{Creates a moodle quiz data dictionary}
\usage{
parse_quiz_metadata(question_xml, question_stats)
}
\arguments{
\item{question_xml}{Output from xml2::read_xml() applied to Moodle Quiz XML file. From the quiz settings, select "Question Bank" then "Export" then download "Moodle XML" file.}

\item{question_stats}{Output from readr::read_csv() applied to Moodle Quiz structure analysis CSV file. From the quiz settings, select "Results" then "Statistics" then download "Quiz structure analysis" table as CSV file.}
}
\value{
Returns a formatted data dictionary for the moodle quiz
}
\description{
Wrangles files directly extracted from moodle to create a formatted data dictionary for the moodle quiz
}
