% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_quiz_response.R
\name{parse_quiz_response}
\alias{parse_quiz_response}
\title{Formats a moodle quiz dataset}
\usage{
parse_quiz_response(response, question_change = NULL)
}
\arguments{
\item{response}{Output from readr::read_csv() applied to Moodle Quiz responses CSV file. From the quiz settings, select "Results" then "Responses" then download table as CSV file.}

\item{question_change}{Default = NULL. Not recommended, but can be used if a question is added after a quiz goes live. Must supply a tibble containing 3 columns: "response" (what question in the dataset that has been added), "change" (the change that has been made - only "new" questions accounted for at present), and "date" (the date the change was made).}
}
\value{
Returns a clean dataset for the moodle quiz
}
\description{
Wrangles files directly extracted from moodle to create a clean dataset for the moodle quiz.
}
